Implement pagnation on Friend request GET API
